Refactor SelfLogBase
