+++ src/OpenTelemetry.Api/SelfLogEventArgs.cs
@@ -1,0 +1,31 @@
+// <copyright file="OpenTelemetryApiEventSource.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+#if XAMARIN
+using System;
+using System.Diagnostics.Tracing;
+
+namespace OpenTelemetry
+{
+    public class SelfLogEventArgs : EventArgs
+    {
+        public int EventId { get; set; }
+        public string EventSource { get; set; }
+        public EventLevel Level { get; set; }
+        public string Message { get; set; }
+    }
+}
+#endif
